requirement.txt
django
pymysql
django-cors-headers
djangorestframework
pyjwt